// src/generated-api/models/FileSummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BMI-LAB Web API
 * BMI-LAB을 관리하기 위한 웹 서비스
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface FileSummary
 */
export interface FileSummary {
  /**
   * 파일 ID (UUID 형식)
   * @type {string}
   * @memberof FileSummary
   */
  fileId?: string;
  /**
   * 파일 이름 (확장자 포함)
   * @type {string}
   * @memberof FileSummary
   */
  fileName?: string;
  /**
   * 파일 사이즈 (bytes 단위)
   * @type {number}
   * @memberof FileSummary
   */
  size?: number;
  /**
   * S3에 업로드된 파일의 URL
   * @type {string}
   * @memberof FileSummary
   */
  uploadUrl?: string;
}

/**
 * Check if a given object implements the FileSummary interface.
 */
export function instanceOfFileSummary(value: object): value is FileSummary {
  return true;
}

export function FileSummaryFromJSON(json: any): FileSummary {
  return FileSummaryFromJSONTyped(json, false);
}

export function FileSummaryFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): FileSummary {
  if (json == null) {
    return json;
  }
  return {
    fileId: json['fileId'] == null ? undefined : json['fileId'],
    fileName: json['fileName'] == null ? undefined : json['fileName'],
    size: json['size'] == null ? undefined : json['size'],
    uploadUrl: json['uploadUrl'] == null ? undefined : json['uploadUrl'],
  };
}

export function FileSummaryToJSON(json: any): FileSummary {
  return FileSummaryToJSONTyped(json, false);
}

export function FileSummaryToJSONTyped(
  value?: FileSummary | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    fileId: value['fileId'],
    fileName: value['fileName'],
    size: value['size'],
    uploadUrl: value['uploadUrl'],
  };
}